# Simplify and optimise proxy header pruning